requirements.txt
aioquic==1.2.0
anyio==4.6.0
asgiref==3.8.1
attr==0.3.2
ConfigParser==7.1.0
contextlib2==21.6.0
ctypes_snappy==1.03
docutils==0.21.2
filelock==3.16.1
flask_swagger_ui==4.11.1
HTMLParser==0.0.2
httpcore==1.0.6
httpx==0.27.2
importlib_metadata==8.0.0
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
kerberos==1.3.1
keyring==25.4.1
mock==5.1.0
protobuf==3.20.2
PyJWT==2.9.0
pymongo_auth_aws==1.3.0
pymongocrypt==1.11.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
railroad==0.5.0
redis==5.1.1
service_identity==24.1.0
sniffio==1.3.1
Sphinx==8.1.3
thread==2.0.5
tornado==6.4.1
trio==0.26.2
urllib3_secure_extra==0.1.0
watchdog==5.0.3
winkerberos==0.12.0
wmi==1.5.1
xmlrpclib==1.0.1
zstandard==0.23.0
